最后一次提交

+++ spzx-service/service-product/src/main/java/com/zq/service/product/service/impl/CategoryServiceImpl.java
@@ -44,19 +44,23 @@ public class CategoryServiceImpl implements CategoryService {
 //            查询全部数据
         List<Category> categoryList = categoryMapper.selectList(null);
 //        挑出一级分类，ParentId=0；
-        List<Category> Onecollect = categoryList.stream().filter(it -> it.getParentId() == 0).toList();
+        List<Category> Onecollect = categoryList.stream().
+                filter(it -> it.getParentId() == 0).toList();
 
-      Onecollect.forEach( one->{
-//        挑出二级分类，并将二级分类添加到以及分类中
-                  List<Category> twocategoryList = categoryList.stream().filter(it -> it.getParentId().longValue() == one.getId()).toList();
+        Onecollect.forEach( one->{
+//        挑出二级分类，并将二级分类添加到一级分类中
+                  List<Category> twocategoryList = categoryList.stream().
+                          filter(it -> it.getParentId() == one.getId()).toList();
+
           one.setChildren(twocategoryList);
 //              跳出三级分类，并将三级分类添加到二级分类中
             twocategoryList.forEach(two ->{
-                List<Category> categoryList3 = categoryList.stream().filter(it -> it.getParentId().longValue() == two.getId()).toList();
+                List<Category> categoryList3 = categoryList.stream().
+                        filter(it -> it.getParentId() == two.getId()).toList();
+
             two.setChildren(categoryList3);
             });
-              }
-       );
+              });
 
 //将一级分类返回
         return Onecollect;
